+++ app/src/test/kotlin/me/xizzhu/android/joshua/search/result/SearchResultPresenterTest.kt
@@ -26,8 +26,6 @@ import me.xizzhu.android.joshua.search.SearchInteractor
 import me.xizzhu.android.joshua.tests.BaseUnitTest
 import me.xizzhu.android.joshua.tests.MockContents
 import me.xizzhu.android.joshua.ui.LoadingSpinnerState
-import me.xizzhu.android.joshua.ui.recyclerview.SearchItem
-import me.xizzhu.android.joshua.ui.recyclerview.TitleItem
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
@@ -92,7 +90,8 @@ class SearchResultPresenterTest : BaseUnitTest() {
     @Test
     fun testObserveDefaultSearchResultAndState() {
         runBlocking {
-            verify(searchResultView, times(1)).onSearchResultUpdated(emptyList())
+            verify(searchResultView, times(1))
+                    .onSearchResultUpdated(emptyList<Verse>().toSearchResult("", searchResultPresenter::selectVerse))
             verify(searchResultView, times(1)).onSearchCompleted()
         }
     }
@@ -106,12 +105,8 @@ class SearchResultPresenterTest : BaseUnitTest() {
             searchResultChannel.send(Pair(query, verses))
             searchStateChannel.send(LoadingSpinnerState.NOT_LOADING)
 
-            verify(searchResultView, times(1)).onSearchResultUpdated(listOf(
-                    TitleItem(MockContents.kjvVerses[0].text.bookName),
-                    SearchItem(MockContents.kjvVerses[0].verseIndex, MockContents.kjvVerses[0].text.bookName, MockContents.kjvVerses[0].text.text, query, searchResultPresenter::selectVerse),
-                    SearchItem(MockContents.kjvVerses[1].verseIndex, MockContents.kjvVerses[1].text.bookName, MockContents.kjvVerses[1].text.text, query, searchResultPresenter::selectVerse),
-                    SearchItem(MockContents.kjvVerses[2].verseIndex, MockContents.kjvVerses[2].text.bookName, MockContents.kjvVerses[2].text.text, query, searchResultPresenter::selectVerse)
-            ))
+            verify(searchResultView, times(1))
+                    .onSearchResultUpdated(verses.toSearchResult(query, searchResultPresenter::selectVerse))
             verify(searchResultView, times(1)).onSearchStarted()
 
             // once from initial state, and second time when search finishes